+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix document properties modal covering only the component area instead of the full screen
+
 ## [0.1.6] - 2026-03-08
 
 ### Fixed