+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.1.7] - 2026-03-08
+
 ### Fixed
 
 - Fix document properties modal covering only the component area instead of the full screen